build: moved server's files to catalog server
